requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
pydantic>=2.8.0
pyyaml>=6.0.1
psycopg[binary,pool]>=3.2.1
pgvector>=0.2.5
sentence-transformers>=3.0.1
google-cloud-aiplatform>=1.66.0
numpy>=1.26.4
pandas>=2.2.2
beautifulsoup4>=4.12.3
trafilatura>=1.7.0
pdfplumber>=0.11.0
tiktoken>=0.7.0
python-dotenv>=1.0.1
